Bayesian estimation of a Markov Random Field of mixed binary and ordinal variables using MCMC.

The function est_bgm explores the joint posterior distribution of the parameters of a Markov Random Field for mixed binary and ordinal variables. It uses a combination of Metropolis-Hastings and Gibbs sampling to create a Markov chain that has the posterior distribution as its invariant distribution. The function uses a pseudolikelihood approach to circumvent having to compute the intractable normalizing constant of the Markov Random Field.

Usage

est_bgm(
  x,
  iter = 10000,
  burnin = 1000,
  interaction_prior = c("Cauchy", "UnitInfo"),
  cauchy_scale = 2.5,
  threshold_alpha = 0.5,
  threshold_beta = 0.5,
  adaptive = FALSE,
  na.action = c("listwise", "impute"),
  save = FALSE,
  display_progress = TRUE
)

Arguments

x

A data frame or matrix with n rows and p columns containing binary and ordinal variables for n independent observations and p variables in the network. Variables are recoded as non-negative integers (0, 1, ..., m) if not already done. Unobserved categories are collapsed into other categories after recoding (i.e., if category 1 is unobserved, the data will be recoded from (0, 2) to (0, 1)).

iter

The number of iterations of the Gibbs sampler. The default of 1e4 is for illustrative purposes. For stable estimates, it is recommended to run the Gibbs sampler for at least 1e5 iterations.

burnin

The number of iterations of the Gibbs sampler before its output is saved. Since it may take some time for the Gibbs sampler to converge to the posterior distribution, it is recommended not to set this number too low.

interaction_prior

The type of prior distribution that is used for the interaction effects. Currently, two prior densities are implemented: The Cauchy prior (interaction_prior = "Cauchy") and the Unit Information prior (interaction_prior = "UnitInfo").

cauchy_scale

The scale of the Cauchy prior for interactions. Defaults to 2.5.

threshold_alpha, threshold_beta

The shape parameters of the beta-prime prior density for the threshold parameters. Must be positive values. If the two values are equal, the prior density is symmetric about zero. If threshold_beta is greater than threshold_alpha, the distribution is skewed to the left, and if threshold_beta is less than threshold_alpha, it is skewed to the right. Smaller values tend to lead to more diffuse prior distributions.

adaptive

Should the function use an adaptive Metropolis algorithm to update interaction parameters within the model? If adaptive = TRUE, bgm adjusts the proposal variance to match the acceptance probability of the random walk Metropolis algorithm to be close to the optimum of .234 using a Robbins-Monro type algorithm. If adaptive = FALSE, it sets the proposal variance to the inverse of the observed Fisher information matrix (the second derivative at the posterior mode). Defaults to FALSE.

na.action

How do you want the function to handle missing data? If na.action = "listwise", listwise deletion is used. If na.action = "impute", missing data are imputed iteratively during the MCMC procedure. Since imputation of missing data can have a negative impact on the convergence speed of the MCMC procedure, it is recommended to run the MCMC for more iterations. Also, since the numerical routines that search for the mode of the posterior do not have an imputation option, the bgm function will automatically switch to interaction_prior = "Cauchy" and adaptive = TRUE.

save

Should the function collect and return all samples from the Gibbs sampler (save = TRUE)? Or should it only return the (model-averaged) posterior means (save = FALSE)? Defaults to FALSE.

display_progress

Should the function show a progress bar (display_progress = TRUE)? Or not (display_progress = FALSE)? Defaults to TRUE.

Value

If save = FALSE (the default), the result is a list of class ``bgms'' containing the following matrices:

interactions: A matrix with p rows and p columns, containing model-averaged posterior means of the pairwise associations.

thresholds: A matrix with p rows and max(m) columns, containing model-averaged category thresholds.

If save = TRUE, the result is a list of class ``bgms'' containing:

interactions: A matrix with iter rows and p * (p - 1) / 2 columns, containing parameter states from every iteration of the Gibbs sampler for the pairwise associations.

thresholds: A matrix with iter rows and sum(m) columns, containing parameter states from every iteration of the Gibbs sampler for the category thresholds.

Column averages of these matrices provide the model-averaged posterior means.

In addition to the analysis results, the bgm output lists some of the arguments of its call. This is useful for post-processing the results.

Details

The prior distribution for the interactions is a Cauchy or Unit-Info type prior. A Beta-prime distribution is used for the exponent of the category parameters.