ADHD symptoms of children aged 6-8 years

A data set containing items measuring symptoms of attention deficit hyperactivity disorder (ADHD) (Silk et al. 2019) Participants are a subsample from a community cohort, the Children's Attention Project (CAP); 355 children, aged 6-8 years, of which 146 were diagnosed with ADHD and 209 were not. ADHD symptoms were assessed during structured face-to-face interviews with the children's parents. The NIMH Diagnostic Interview Schedule for Children IV-DISC-iV (Shaffer et al. 2000) contains a total of 18 symptoms, of which 9 Inattentive and 9 Hyperactive/Impulsive. Each symptom was scored as either present (1) or absent (0), based on the parents' statements.

Usage

data("ADHD")

Format

A matrix with 355 rows and 19 columns. (I) and (HI) indicate Inattentive and Hyperactive/Impulsive symptoms, respectively.

group

1 = ADHD diagnosis, 0 = no ADHD diagnosis

avoid

Often avoids, dislikes, or is reluctant to engage in tasks that require sustained mental effort (I)

closeatt

Often fails to give close attention to details or makes careless mistakes in schoolwork, work, or other activities (I)

distract

Is often easily distracted by extraneous stimuli (I)

forget

Is often forgetful in daily activities (I)

instruct

Often does not follow through on instructions and fails to finish schoolwork, chores, or duties in the workplace (I)

listen

Often does not seem to listen when spoken to directly (I)

loses

Often loses things necessary for tasks or activities (I)

org

Often has difficulty organizing tasks and activities (I)

susatt

Often has difficulty sustaining attention in tasks or play activities (I)

blurts

Often blurts out answers before questions have been completed (HI)

fidget

Often fidgets with hands or feet or squirms in seat (HI)

interrupt

Often interrupts or intrudes on others (HI)

motor

Is often "on the go" or often acts as if "driven by a motor" (HI)

quiet

Often has difficulty playing or engaging in leisure activities quietly (HI)

runs

Often runs about or climbs excessively in situations in which it is inappropriate (HI)

seat

Often leaves seat in classroom or in other situations in which remaining seated is expected (HI)

talks

Often talks excessively (HI)

turn

Often has difficulty awaiting turn (HI)

Source

Silk et al. (2019) . Data retrieved from https://doi.org/10.1371/journal.pone.0211053.s004. Licensed under the Creative Commons Attribution 4.0 International License (CC-BY 4.0). https://creativecommons.org/licenses/by/4.0/

References

Shaffer D, Fisher P, Lucas CP, Dulcan MK, Schwab-Stone ME (2000). “NIMH Diagnostic Interview Schedule for Children Version IV (NIMH DISC-IV): description, differences from previous versions, and reliability of some common diagnoses.” Journal of the American Academy of Child & Adolescent Psychiatry, 39, 28–38. doi:10.1097/00004583-200001000-00014 , PMID: 10638065.

Silk TJ, Malpas CB, Beare R, Efron D, Anderson V, Hazell P, Jongeling B, Nicholson JM, Sciberras E (2019). “A network analysis approach to ADHD symptoms: More than the sum of its parts.” PLOS ONE, 14(1), e0211053. doi:10.1371/journal.pone.0211053 .