bgms: Bayesian Analysis of Networks of Binary and/or Ordinal Variables

The R package bgms provides tools for Bayesian analysis of graphical models describing networks of variables. The package uses Markov chain Monte Carlo methods combined with a pseudolikelihood approach to estimate the posterior distribution of model parameters.

Gibbs variable selection (George and McCulloch 1993) is used to model the underlying network structure of the graphical model. By imposing a discrete spike and slab prior on the pairwise interactions, it is possible to shrink the interactions to exactly zero. The Gibbs sampler embeds a Metropolis approach for mixtures of mutually singular distributions (Gottardo and Raftery 2008) to account for the discontinuity at zero. The goal is to provide these tools for Markov Random Field (MRF) models for a wide range of variable types in the bgms package, and it currently provides them for analyzing networks of binary and/or ordinal variables (Marsman and Haslbeck 2023) .

While the goal is to provide the above tools for Markov Random Field (MRF) models for a wide range of variable types in the bgms package, it currently provides tools for analyzing networks of binary and/or ordinal variables (Marsman and Haslbeck 2023) .

MRFs are a special class of graphical models whose graph structure reflects the conditional associations between their variables, making them useful for testing for conditional independence or dependence. For example, the inclusion Bayes factor tests for conditional independence or dependence of a pair of variables in the network by comparing the predictive adequacy of models that include the edge between these variables and models that exclude the edge. (Huth et al. 2023; Sekulovski et al. 2023) .

The bgms package offers several tools for analyzing the structure of the MRF:

Simulate response data from the MRF using the Gibbs sampler.

Simulate mrfSampler.

Estimate the posterior distribution of the MRF's parameters and possibly its network structure using Gibbs variable selection.

Bayesian estimation or Bayesian edge selection with bgm.

References

George EI, McCulloch RE (1993). “Variable selection via Gibbs sampling.” Journal of the American Statistical Association, 88(423), 881-889. doi:10.1080/01621459.1993.10476353 .

Gottardo R, Raftery AE (2008). “Markov Chain Monte Carlo With Mixtures of Mutually Singular Distributions.” Journal of Computational and Graphical Statistics, 17(4), 949--975. doi:10.1198/106186008X386102 .

Huth K, de Ron J, Goudriaan AE, Luigjes K, Mohammadi R, van Holst RJ, Wagenmakers E, Marsman M (2023). “Bayesian analysis of cross-sectional networks: A tutorial in R and JASP.” Advances in Methods and Practices in Psychological Science, 6(4), 1-18. doi:10.1177/25152459231193334 .

Marsman M, Haslbeck JMB (2023). “Bayesian analysis of the ordinal Markov Random Field.” PsyArXiv. doi:10.31234/osf.io/ukwrf .

Sekulovski N, Keetelaar S, Huth K, Wagenmakers E, van Bork R, van den Bergh D, Marsman M (2023). “Testing conditional independence in psychometric networks: An analysis of three Bayesian methods.” PsyArXiv. doi:10.31234/osf.io/ch7a2 .

See also

Author

Maintainer: Maarten Marsman m.marsman@uva.nl (ORCID)

Other contributors:

Karoline Huth (ORCID) [contributor]

Nikola Sekulovski (ORCID) [contributor]

Don van den Bergh (ORCID) [contributor]