bgms: Bayesian Analysis of Networks of Binary and/or Ordinal Variables

The R package bgms provides tools for Bayesian analysis of the ordinal Markov random field, a graphical model describing a network of binary and/or ordinal variables MarsmanVandenBerghHaslbeck_2024bgms. A pseudolikelihood is used to approximate the likelihood of the graphical model, and Markov chain Monte Carlo methods are used to simulate from the corresponding pseudoposterior distribution of the graphical model parameters.

The bgm function can be used for a one-sample design and the bgmCompare function can be used for a two-independent-samples design MarsmanWaldorpSekulovskiHaslbeck_2024bgms. Both functions can model the selection of effects. In one-sample designs, the bgm function models the presence or absence of edges between pairs of variables in the network. The estimated posterior inclusion probability indicates how plausible it is that a network with an edge between the two corresponding variables produced the observed data, and can be converted into a Bayes factor test for conditional independence.

In two-independent-samples designs, the bgmCompare function models the selection of group differences in edge weights and possibly category thresholds. The estimated posterior inclusion probability indicates how plausible it is that graphical models with a difference in the corresponding edge weight or category threshold generated the data at hand, and can be converted to a Bayes factor test for parameter equivalence.

The bgms package offers several tools for analyzing the structure of the MRF:

Simulate response data from the MRF using the Gibbs sampler.

Simulate mrfSampler.

Estimate the posterior distribution of the MRF's parameters and possibly its network structure in one-sample designs.

Bayesian estimation or Bayesian edge selection with bgm.

Estimate the posterior distribution of the MRF's parameters in a two-independent-sample design, and possibly perform selection on group differences in MRF parameters.

Bayesian estimation or Bayesian difference selection with bgmCompare.

References

See also

Author

Maintainer: Maarten Marsman m.marsman@uva.nl (ORCID)

Other contributors:

Giuseppe Arena (ORCID) [contributor]

Karoline Huth (ORCID) [contributor]

Nikola Sekulovski (ORCID) [contributor]

Don van den Bergh (ORCID) [contributor]