Optimization of the Pseudoposterior for a Markov Random Field model for ordinal variables.

The function mppe estimates the parameters for the ordinal MRF by optimizing the pseudoposterior with Newton-Raphson.

Usage

mppe(
  x,
  no_categories,
  interaction_prior = "Cauchy",
  cauchy_scale = 2.5,
  threshold_alpha = 1,
  threshold_beta = 1,
  convergence_criterion = sqrt(.Machine$double.eps),
  maximum_iterations = 1000,
  thresholds,
  interactions
)

Arguments

x

An no_persons by no_nodes matrix containing the categories coded as non-negative integers (i.e., coded 0, 1, ..., no_categories) for no_persons independent observations on no_nodes variables in the network or graph.

no_categories

The maximum category.

interaction_prior

The prior distribution for the interaction effects. Currently, two prior densities are implemented: The Unit Information prior (interaction_prior = "UnitInfo") and the Cauchy prior (interaction_prior = "Cauchy"). Defaults to "Cauchy".

cauchy_scale

The scale of the Cauchy prior for interactions. Defaults to 2.5.

threshold_alpha, threshold_beta

The shape parameters of the Beta-prime prior for the thresholds. Default to 1.

convergence_criterion

The convergence criterion for the pseudoposterior values in the EM algorithm. Defaults to sqrt(.Machine$double.eps).

maximum_iterations

The maximum number of EM iterations used. Defaults to 1e3. A warning is issued if procedure has not converged in maximum_iterations iterations.

thresholds

A no_nodes by no_categories matrix thresholds. Used as starting values in the Newton-Raphson procedure. Optional.

interactions

A no_nodes by no_nodes matrices interactions. Used as starting values in the Newton-Raphson procedure. Optional.

Value

A list containing the no_nodes by no_nodes matrices interactions, the no_nodes by no_categories matrix thresholds, the hessian matrix as computed in the final iteration of the optimization procedure, and, in case interaction_prior = "UnitInfo", the no_nodes by no_nodes

matrix unit_info. The matrix interactions is a numeric matrix that contains the maximum pseudoposterior estimates of the pairwise associations. The matrix thresholds contains the maximum pseudoposterior estimates of the category thresholds parameters. The hessian matrix has dimensions equal to the number of thresholds + associations. The topleft square contains the thresholds, the bottomright square the associations (of the form (1,2), (1, 3), ..., (2, 1), ...). The unit_information matrix contains the variances of the unit information prior for the association effects.